pretty much done?

+++ src/controller/PokeTeamController.ts
@@ -5,10 +5,12 @@ import { PokeTeam } from '../entity/PokeTeam'
 import { Controller } from '../decorator/Controller'
 import { Route } from '../decorator/Route'
 import { validate, ValidationError, ValidatorOptions } from 'class-validator'
+import { User } from '../entity/User'
 
 @Controller('/poketeam')
 export default class PokeTeamController {
   private readonly pokeTeamRepo = AppDataSource.getRepository(PokeTeam)
+  private readonly userRepo = AppDataSource.getRepository(User)
 
   // https://github.com/typestack/class-validator#passing-options
   private readonly validOptions: ValidatorOptions = {
@@ -19,6 +21,24 @@ export default class PokeTeamController {
 
   @Route('get', '/:teamID*?') // the *? makes the param optional - see https://expressjs.com/en/guide/routing.html#route-paramters
   async read (req: Request, res: Response, next: NextFunction): Promise<PokeTeam | PokeTeam[]> {
+    const token = req.header('Authorization')?.replace('Bearer ', '')
+    console.log('This is the token ' + token)
+    let user = new User()
+    if (token) {
+      user = await this.userRepo.findOneBy({ token })
+      if (!user) {
+        return res.status(401).json({ error: 'token does not exist' })
+      }
+    } else {
+      return res.status(401).json({ error: 'token does not exist' })
+    }
+    const accessLevel = user.accessLevel
+    const rType = req.method
+
+    console.log('This is the user: ' + user.username)
+    console.log('This is the method being requested: ' + req.method)
+    console.log("This is the user's access level: " + user.accessLevel)
+
     if (req.params.teamID) return await this.pokeTeamRepo.findOneBy({ teamID: req.params.teamID })
     else {
       const findOptions: any = { order: {} } // prepare order and where props
@@ -43,6 +63,27 @@ export default class PokeTeamController {
 
   @Route('delete', '/:teamID')
   async delete (req: Request, res: Response, next: NextFunction): Promise<PokeTeam> {
+    const token = req.header('Authorization')?.replace('Bearer ', '')
+    console.log('This is the token ' + token)
+    let user = new User()
+    if (token) { // re-assign based on token
+      user = await this.userRepo.findOneBy({ token })
+      if (!user) {
+        return res.status(401).json({ error: 'token does not exist' })
+      }
+    } else {
+      return res.status(401).json({ error: 'token does not exist' })
+    }
+    const accessLevel = user.accessLevel
+    const rType = req.method
+    console.log('This is the user: ' + user.username)
+    console.log('This is the method being requested: ' + req.method)
+    console.log("This is the user's access level: " + user.accessLevel)
+
+    if (accessLevel !== 'ADMIN') {
+      return res.status(401).json({ error: 'user does not have permission' })
+    }
+
     const pokeTeamToRemove = await this.pokeTeamRepo.findOneBy({ teamID: req.params.teamID })
     // res.statusCode = 204 --No Content - browser will complain since we are actually returning content
     if (pokeTeamToRemove) return await this.pokeTeamRepo.remove(pokeTeamToRemove)
@@ -51,6 +92,27 @@ export default class PokeTeamController {
 
   @Route('post')
   async create (req: Request, res: Response, next: NextFunction): Promise<PokeTeam | ValidationError[]> {
+    const token = req.header('Authorization')?.replace('Bearer ', '')
+    console.log('This is the token ' + token)
+    let user = new User()
+    if (token) { // re-assign based on token
+      user = await this.userRepo.findOneBy({ token })
+      if (!user) {
+        return res.status(401).json({ error: 'token does not exist' })
+      }
+    } else {
+      return res.status(401).json({ error: 'token does not exist' })
+    }
+    const accessLevel = user.accessLevel
+    const rType = req.method
+    console.log('This is the user: ' + user.username)
+    console.log('This is the method being requested: ' + req.method)
+    console.log("This is the user's access level: " + user.accessLevel)
+
+    if (!(accessLevel === 'ADMIN' || accessLevel === 'WRITE')) {
+      return res.status(401).json({ error: 'user does not have permission' })
+    }
+
     const newPokeTeam = Object.assign(new PokeTeam(), req.body)
     const violations = await validate(newPokeTeam, this.validOptions)
     if (violations.length) {
@@ -63,6 +125,26 @@ export default class PokeTeamController {
 
   @Route('put', '/:teamID')
   async update (req: Request, res: Response, next: NextFunction): Promise<PokeTeam | ValidationError[]> {
+    const token = req.header('Authorization')?.replace('Bearer ', '')
+    let user = new User()
+    if (token) { // re-assign based on token
+      user = await this.userRepo.findOneBy({ token })
+      if (!user) {
+        return res.status(401).json({ error: 'token does not exist' })
+      }
+    } else {
+      return res.status(401).json({ error: 'token does not exist' })
+    }
+    const accessLevel = user.accessLevel
+    const rType = req.method
+    console.log('This is the user: ' + user.username)
+    console.log('This is the method being requested: ' + req.method)
+    console.log("This is the user's access level: " + user.accessLevel)
+
+    if (!(accessLevel === 'ADMIN' || accessLevel === 'WRITE')) {
+      return res.status(401).json({ error: 'user does not have permission' })
+    }
+
     const pokeTeamToUpdate = await this.pokeTeamRepo.preload(req.body)
     // Extra validation - ensure the id param matched the id submitted in the body
     if (!pokeTeamToUpdate || pokeTeamToUpdate.teamID.toString() !== req.params.teamID) next() // pass the buck until 404 error is sent
